mosquitto: bump to 1.4-development branch
This is an unreleased preview image.

libwebsockets support is included and enabled by default.  You will need
to build the package yourself to disable this, it's not considered a
variant useful enough to have it's own binaries built.

Note that libwebsockets-openssl is chosen at present, as that's the ssl
library supported by mosquitto.

If you really don't want web sockets support, install the -nossl
variant, or rebuild mosquitto yourself, deselecting the websockes
support option.